Fix decoder visibility, integer overflow crash, and Float fidelity
- Make DictionaryDecoder.init() public so external modules can instantiate it
- Throw EncodingError instead of trapping when an integer (e.g. UInt64.max)
  does not fit in Int
- Store Float natively in DictionaryValue instead of widening to Double, so it
  round-trips with its own precision and type
- Expand test coverage from 5 to 35 cases (round-trip, error handling, all
  date/data/key strategies, super encoder/decoder)
- Document integer range limit, Float support, and snake_case acronym behavior

+++ Tests/DictionaryCoderTests/DictionaryEncoderTests.swift
@@ -18,7 +18,7 @@ final class DictionaryEncoderTests: XCTestCase {
         XCTAssertEqual(dictionary["int"] as! Int, Int(64))
         XCTAssertEqual(dictionary["int8"] as! Int, Int(8))
         XCTAssertEqual(dictionary["double"] as! Double, Double(0.5))
-        XCTAssertEqual(dictionary["float"] as! Double, Double(0.5))
+        XCTAssertEqual(dictionary["float"] as! Float, Float(0.5))
         XCTAssertEqual(dictionary["decimal"] as! Decimal, Decimal(0.1))
         XCTAssertEqual(dictionary["string"] as! String, "String")
         XCTAssertEqual(dictionary["array"] as! [String], ["String0", "String1"])
